test(client): share Last Will broker coverage
